Refactor storage, fix zappdev
