+++ src/client/storage/init/init.ts
@@ -1,0 +1,9 @@
+import {uploadLoop} from '../uploader';
+
+const init = () => {
+    uploadLoop();
+    // TODO: re-enable this validation!
+    // chunkValidatorLoop();
+};
+
+export default init;
