+++ src/client/storage/init/init_arlocal.ts
@@ -1,0 +1,13 @@
+import axios from 'axios';
+import {HOST, PORT, PROTOCOL} from '../config';
+import arweave, {arweaveKey} from '../arweave/arweave_arlocal';
+
+const init = async () => {
+    // mint tokens on arlocal
+    if (arweaveKey !== undefined) {
+        const address = await arweave.wallets.jwkToAddress(arweaveKey);
+        await axios.get(`${PROTOCOL}://${HOST}:${PORT}/mint/${address}/100000000000000000000`);
+    }
+};
+
+export default init;
